# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CineTrack API",
    "version": "1.0.0",
    "description": "OpenAPI specification for the CineTrack Fastify backend."
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/v1",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/dashboard/executive": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Get executive dashboard summary",
        "parameters": [
          {
            "$ref": "#/components/parameters/StartDateParam"
          },
          {
            "$ref": "#/components/parameters/EndDateParam"
          },
          {
            "$ref": "#/components/parameters/CityParam"
          },
          {
            "$ref": "#/components/parameters/CinemaIdQueryParam"
          },
          {
            "name": "studio_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Executive dashboard",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                },
                "example": {
                  "success": true,
                  "data": {
                    "system_status": {
                      "status": "active",
                      "last_update": "2026-05-14T12:00:00.000Z",
                      "health_rate": 100
                    },
                    "active_vs_inactive_cinema": {
                      "total": 20,
                      "active": 20,
                      "inactive": 0,
                      "maintenance": 0,
                      "down": 0,
                      "is_mock_maintenance": true
                    },
                    "franchise_performance": [],
                    "city_revenue_summary": [],
                    "top_film_contribution": [],
                    "alert_summary": {
                      "critical": 0,
                      "warning": 0,
                      "info": 2,
                      "unread": 0
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/sales/overview": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Get sales overview metrics",
        "parameters": [
          {
            "$ref": "#/components/parameters/StartDateParam"
          },
          {
            "$ref": "#/components/parameters/EndDateParam"
          },
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "daily",
                "weekly",
                "monthly"
              ]
            }
          },
          {
            "$ref": "#/components/parameters/CityParam"
          },
          {
            "$ref": "#/components/parameters/CinemaIdQueryParam"
          },
          {
            "name": "studio_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sales overview",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                },
                "example": {
                  "success": true,
                  "data": {
                    "total_revenue": 318897536,
                    "total_tickets": 7847,
                    "avg_ticket_price": 40639.47,
                    "revenue_per_seat": 8858.27,
                    "avg_occupancy": 19.81,
                    "growth": {
                      "tickets": 10.01,
                      "revenue": 17.6,
                      "avg_occupancy": 1.59
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/sales/revenue-by-cinema": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Get revenue breakdown by cinema",
        "parameters": [
          {
            "$ref": "#/components/parameters/StartDateParam"
          },
          {
            "$ref": "#/components/parameters/EndDateParam"
          },
          {
            "$ref": "#/components/parameters/CityParam"
          },
          {
            "$ref": "#/components/parameters/CinemaIdQueryParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Revenue by cinema",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                },
                "example": {
                  "success": true,
                  "data": {
                    "summary": {
                      "total_cinemas": 20,
                      "active_cinemas": 20
                    },
                    "top_performing_cinema": {
                      "rank": 1,
                      "cinema_id": "C001",
                      "cinema_name": "Semarang Mall 1 XXI",
                      "city": "Semarang",
                      "total_revenue": 12500000,
                      "total_tickets": 280,
                      "contribution": 6.5,
                      "top_movie": null
                    },
                    "lowest_performing_cinema": null,
                    "breakdown": []
                  }
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/sales/revenue-by-movie": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Get revenue breakdown by movie",
        "parameters": [
          {
            "$ref": "#/components/parameters/StartDateParam"
          },
          {
            "$ref": "#/components/parameters/EndDateParam"
          },
          {
            "$ref": "#/components/parameters/CityParam"
          },
          {
            "$ref": "#/components/parameters/CinemaIdQueryParam"
          },
          {
            "name": "top_n",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Revenue by movie",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                },
                "example": {
                  "success": true,
                  "data": {
                    "total_movies": 8,
                    "top_movie": {
                      "rank": 1,
                      "movie_id": "M005",
                      "title": "Siksa Kubur",
                      "total_revenue": 132880000,
                      "total_tickets": 2516,
                      "contribution": 17.2
                    },
                    "breakdown": []
                  }
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/sales/time-slots": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Get sales metrics by time slot",
        "parameters": [
          {
            "$ref": "#/components/parameters/StartDateParam"
          },
          {
            "$ref": "#/components/parameters/EndDateParam"
          },
          {
            "$ref": "#/components/parameters/CityParam"
          },
          {
            "$ref": "#/components/parameters/CinemaIdQueryParam"
          },
          {
            "name": "studio_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "movie_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sales time slots",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                },
                "example": {
                  "success": true,
                  "data": {
                    "peak_sales_hour": {
                      "time_slot": "18:00",
                      "revenue": 210605000,
                      "demand": 3976,
                      "occupancy": 44.18,
                      "normalized_demand": 1,
                      "normalized_revenue": 1,
                      "optimization_score": 0.2,
                      "recommendation": "Pertahankan slot ini sambil memantau perubahan demand."
                    },
                    "quiet_hour": {
                      "time_slot": "15:00",
                      "revenue": 151545000,
                      "demand": 2859,
                      "occupancy": 31.77,
                      "normalized_demand": 0,
                      "normalized_revenue": 0,
                      "optimization_score": 0
                    },
                    "breakdown": []
                  }
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/sales/trend": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Get sales trend with moving average and forecast",
        "parameters": [
          {
            "$ref": "#/components/parameters/StartDateParam"
          },
          {
            "$ref": "#/components/parameters/EndDateParam"
          },
          {
            "name": "group_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "daily",
                "monthly"
              ]
            }
          },
          {
            "$ref": "#/components/parameters/CityParam"
          },
          {
            "$ref": "#/components/parameters/CinemaIdQueryParam"
          },
          {
            "name": "studio_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "movie_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sales trend",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                },
                "example": {
                  "success": true,
                  "data": {
                    "summary": {
                      "group_by": "daily",
                      "total_revenue": 318897536,
                      "total_tickets": 7847
                    },
                    "forecast": {
                      "group_by": "daily",
                      "next_period": "next_period",
                      "projected_revenue": 103000000,
                      "is_mock_forecast": true
                    },
                    "breakdown": []
                  }
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/sales/payment": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Get payment preference and profitability",
        "parameters": [
          {
            "$ref": "#/components/parameters/StartDateParam"
          },
          {
            "$ref": "#/components/parameters/EndDateParam"
          },
          {
            "$ref": "#/components/parameters/CityParam"
          },
          {
            "$ref": "#/components/parameters/CinemaIdQueryParam"
          },
          {
            "name": "studio_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "payment_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sales payment metrics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                },
                "example": {
                  "success": true,
                  "data": {
                    "total_transactions": 7847,
                    "preferred_payment": {
                      "payment_type": "QRIS",
                      "total_transactions": 3500,
                      "usage_rate": 44.6,
                      "gross_revenue": 145000000,
                      "avg_price": 41428.57,
                      "admin_fee": 1000,
                      "success_rate": 100,
                      "failure_rate": 0,
                      "net_profitability": 141500000,
                      "is_mock_config": true
                    },
                    "breakdown": []
                  }
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/sales/operational-risk": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Get sales operational risk summary",
        "parameters": [
          {
            "$ref": "#/components/parameters/StartDateParam"
          },
          {
            "$ref": "#/components/parameters/EndDateParam"
          },
          {
            "$ref": "#/components/parameters/CityParam"
          },
          {
            "$ref": "#/components/parameters/CinemaIdQueryParam"
          },
          {
            "name": "studio_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "movie_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sales operational risk",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                },
                "example": {
                  "success": true,
                  "data": {
                    "summary": {
                      "total_shows": 100,
                      "cancelled": 2,
                      "delayed": 1,
                      "avg_delay_minutes": null,
                      "problematic_rate": 3,
                      "is_mock_delay": true
                    },
                    "problematic_schedules": []
                  }
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/films/overview": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Get films overview metrics",
        "parameters": [
          {
            "$ref": "#/components/parameters/StartDateParam"
          },
          {
            "$ref": "#/components/parameters/EndDateParam"
          },
          {
            "$ref": "#/components/parameters/CityParam"
          },
          {
            "$ref": "#/components/parameters/CinemaIdQueryParam"
          },
          {
            "name": "studio_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Films overview",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                },
                "example": {
                  "success": true,
                  "data": {
                    "active_films": 8,
                    "total_shows": 120,
                    "tickets_sold": 7847,
                    "revenue": 318897536,
                    "avg_occupancy": 19.81,
                    "growth": {
                      "tickets": 10.01,
                      "revenue": 17.6,
                      "avg_occupancy": 1.59
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/films/performance": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Get films performance and blockbuster ranking",
        "parameters": [
          {
            "$ref": "#/components/parameters/StartDateParam"
          },
          {
            "$ref": "#/components/parameters/EndDateParam"
          },
          {
            "$ref": "#/components/parameters/CityParam"
          },
          {
            "$ref": "#/components/parameters/CinemaIdQueryParam"
          },
          {
            "name": "top_n",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Films performance",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                },
                "example": {
                  "success": true,
                  "data": {
                    "top_movie": {
                      "rank": 1,
                      "movie_id": "M005",
                      "title": "Siksa Kubur",
                      "total_tickets": 2516,
                      "total_revenue": 132880000,
                      "blockbuster_score": 54628307.2
                    },
                    "breakdown": []
                  }
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/films/schedules": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Get film schedule performance and audience density",
        "parameters": [
          {
            "$ref": "#/components/parameters/StartDateParam"
          },
          {
            "$ref": "#/components/parameters/EndDateParam"
          },
          {
            "$ref": "#/components/parameters/CityParam"
          },
          {
            "$ref": "#/components/parameters/CinemaIdQueryParam"
          },
          {
            "name": "studio_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "movie_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Film schedules analytics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                },
                "example": {
                  "success": true,
                  "data": {
                    "schedule_performance": [],
                    "repeat_schedule_performance": [],
                    "audience_density": []
                  }
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/films/occupancy": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Get films occupancy breakdown",
        "parameters": [
          {
            "$ref": "#/components/parameters/StartDateParam"
          },
          {
            "$ref": "#/components/parameters/EndDateParam"
          },
          {
            "$ref": "#/components/parameters/CityParam"
          },
          {
            "$ref": "#/components/parameters/CinemaIdQueryParam"
          },
          {
            "name": "studio_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "movie_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Films occupancy",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                },
                "example": {
                  "success": true,
                  "data": {
                    "overall": {
                      "group_by": "daily",
                      "total_tickets": 7847,
                      "total_capacity": 36000,
                      "occupancy": 19.81
                    },
                    "by_movie": [],
                    "by_day": [],
                    "by_studio": []
                  }
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/films/distribution": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Get genre popularity and studio format distribution",
        "parameters": [
          {
            "$ref": "#/components/parameters/StartDateParam"
          },
          {
            "$ref": "#/components/parameters/EndDateParam"
          },
          {
            "$ref": "#/components/parameters/CityParam"
          },
          {
            "$ref": "#/components/parameters/CinemaIdQueryParam"
          },
          {
            "name": "studio_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Films distribution",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                },
                "example": {
                  "success": true,
                  "data": {
                    "genre_popularity": {
                      "summary": {
                        "total_movies_showing": 8,
                        "total_tickets_sold": 7847,
                        "top_movie": null,
                        "top_genre": null
                      },
                      "breakdown": [],
                      "breakdown_rating_usia": []
                    },
                    "studio_format_distribution": []
                  }
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/films/operational-risk": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Get film operational risk summary",
        "parameters": [
          {
            "$ref": "#/components/parameters/StartDateParam"
          },
          {
            "$ref": "#/components/parameters/EndDateParam"
          },
          {
            "$ref": "#/components/parameters/CityParam"
          },
          {
            "$ref": "#/components/parameters/CinemaIdQueryParam"
          },
          {
            "name": "studio_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "movie_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Films operational risk",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                },
                "example": {
                  "success": true,
                  "data": {
                    "summary": {
                      "total_shows": 100,
                      "cancelled": 2,
                      "delayed": 1,
                      "avg_delay_minutes": null,
                      "problematic_rate": 3,
                      "is_mock_delay": true
                    },
                    "impacted_movies": [],
                    "problematic_schedules": []
                  }
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/notifications": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "List dashboard notifications",
        "parameters": [
          {
            "$ref": "#/components/parameters/PageParam"
          },
          {
            "$ref": "#/components/parameters/LimitParam"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "read",
                "unread",
                "all"
              ]
            }
          },
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "info",
                "warning",
                "critical",
                "success"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dashboard notifications",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                },
                "example": {
                  "success": true,
                  "data": [
                    {
                      "type": "system-health",
                      "severity": "success",
                      "title": "Backend health",
                      "what_happened": "Backend is active and responding normally.",
                      "where": "network",
                      "impact_size": 334677,
                      "recommended_action": "Cek konektivitas backend, database, dan sinkronisasi data.",
                      "created_at": "2026-04-10T08:15:33.194Z",
                      "status": "read",
                      "resolved": true
                    }
                  ],
                  "meta": {
                    "filters": {
                      "status": "all",
                      "severity": null
                    },
                    "pagination": {
                      "page": 1,
                      "limit": 20,
                      "total": 4,
                      "total_pages": 1
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
    "parameters": {
      "PageParam": {
        "name": "page",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1
        },
        "required": false
      },
      "LimitParam": {
        "name": "limit",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1
        },
        "required": false
      },
      "StartDateParam": {
        "name": "start_date",
        "in": "query",
        "schema": {
          "type": "string",
          "format": "date"
        },
        "required": false
      },
      "EndDateParam": {
        "name": "end_date",
        "in": "query",
        "schema": {
          "type": "string",
          "format": "date"
        },
        "required": false
      },
      "CityParam": {
        "name": "city",
        "in": "query",
        "schema": {
          "type": "string"
        },
        "required": false
      },
      "CinemaIdQueryParam": {
        "name": "cinema_id",
        "in": "query",
        "schema": {
          "type": "string"
        },
        "required": false
      },
...
    },
    "schemas": {
      "SuccessEnvelope": {
        "type": "object",
        "required": [
          "success",
          "data"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "const": true
          },
          "data": {},
          "meta": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
...
    }
  }
}